extracted drag behaviors to their own class
